Update: heart mesh는 heart.js 파일로 분리"
