+++ src/script.js
@@ -1,6 +1,7 @@
 import './style.css'
 import * as THREE from 'three'
 import * as dat from 'dat.gui'
+import {heartMesh} from './heart';
 
 // Debug
 const gui = new dat.GUI()
@@ -11,27 +12,7 @@ const canvas = document.querySelector('#webgl')
 // Scene
 const scene = new THREE.Scene()
 
-// Geometry
-const heartShape = new THREE.Shape();
-
-heartShape.moveTo( 25, 25 );
-heartShape.bezierCurveTo( 25, 25, 20, 0, 0, 0 );
-heartShape.bezierCurveTo( - 30, 0, - 30, 35, - 30, 35 );
-heartShape.bezierCurveTo( - 30, 55, - 10, 77, 25, 95 );
-heartShape.bezierCurveTo( 60, 77, 80, 55, 80, 35 );
-heartShape.bezierCurveTo( 80, 35, 80, 0, 50, 0 );
-heartShape.bezierCurveTo( 35, 0, 25, 25, 25, 25 );
-
-const extrudeSettings = { depth: 20, bevelEnabled: true, bevelSegments: 2, steps: 2, bevelSize: 2, bevelThickness: 0.5 };
-const geometry = new THREE.ExtrudeGeometry( heartShape, extrudeSettings );
-
-// Material
-const material = new THREE.MeshBasicMaterial({color: 0xFF1493, wireframe: true})
-
 // Mesh
-const heartMesh = new THREE.Mesh( geometry, material);
-heartMesh.geometry.rotateX(Math.PI)
-heartMesh.geometry.translate(-25, 100, 0)
 scene.add(heartMesh)
 
 // Lights
@@ -76,10 +57,10 @@ const clock = new THREE.Clock()
 const tick = () =>
 {
     // Update objects
-    const elapsedTime = clock.getElapsedTime()
-    heartMesh.rotation.x = 1.3 * elapsedTime
-    heartMesh.rotation.y = 1.5 * elapsedTime
-    heartMesh.rotation.z = 1.2 * elapsedTime
+    // const elapsedTime = clock.getElapsedTime()
+    // heartMesh.rotation.x = 1.3 * elapsedTime
+    // heartMesh.rotation.y = 1.5 * elapsedTime
+    // heartMesh.rotation.z = 1.2 * elapsedTime
     
     // Render
     renderer.render(scene, camera)
